// fabric-frontend/apis/define/Subject.ts
export interface Subject {
    id: string;
    institution: string;
    leader: string;
    leader_contact_method: string;
    name: string;
    project_id: string;
    created_time: string;
    members_count: string;
}

export interface subject{
    id:string;
    institution:string;
    leader:string;
    name:string;
}